requirements.txt
numpy==1.24.3
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
PyYAML==6.0.1
scipy==1.11.3
matplotlib==3.8.0
pandas==2.1.1
jupyter==1.0.0
ipywidgets==8.1.1
pytest==7.4.2
pytest-cov==4.1.0
tqdm==4.66.1
pillow==10.0.1